requirements.txt
fastapi
uvicorn
gunicorn
redis
openai
python-dotenv
gtts
openai-whisper